backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
openai==1.65.3
gunicorn==23.0.0
python-dotenv
requests